Refactoring: Moved code related to socket connection into separate class.
