+++ iai_kms_40_driver/CMakeLists.txt
@@ -97,6 +97,7 @@ include_directories(
 )
 
 add_library(${PROJECT_NAME}
+  src/${PROJECT_NAME}/socket_connection.cpp
   src/${PROJECT_NAME}/driver.cpp)
 target_link_libraries(${PROJECT_NAME}
   ${catkin_LIBRARIES})
